@@ -2,7 +2,7 @@ import { NgModule, ModuleWithProviders } from '@angular/core';
 
 @NgModule()
 export class BrowserLoggerModule {
-  public static forRoot(environment: any): ModuleWithProviders {
+  public static forRoot(environment: any): ModuleWithProviders<BrowserLoggerModule> {
     return {
       ngModule: BrowserLoggerModule,
       providers: [
